lambda/diseasedetection/requirements.txt
tensorflow-cpu
Pillow